Fixed greedy solver. And added numba jit to score function.

+++ src/arc2020/mytypes.py
@@ -6,7 +6,7 @@ import typing as t
 class ImgMatrix(np.ndarray):
 
     def __new__(cls, *args):
-        self = np.array(*args).view(cls)
+        self = np.array(*args, dtype=np.uint8).view(cls)
         return self
